+++ src/wrapper.rs
@@ -13,15 +13,19 @@ mod option;
 mod default_val;
 mod default_trait;
 mod vec;
+mod vec_deque;
 mod map;
 
 pub use option::OptionValue;
 pub use default_val::DefaultValue;
 pub use default_trait::DefaultTrait;
 pub use map::Map;
+pub use vec_deque::VecDeque;
 pub use vec::Vec;
 
 // we need to expose prefixed for the generated 
 // code to be able to access it
 #[doc(hidden)]
-pub use vec::Prefixed;
+pub use vec::Prefixed as VecPrefixed;
+#[doc(hidden)]
+pub use vec_deque::Prefixed as DequePrefixed;